% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_genbank}
\alias{read_genbank}
\title{Read a GenBank file}
\usage{
read_genbank(file)
}
\arguments{
\item{file}{Path to a GenBank file.}
}
\value{
Parsed records in legacy list format.
}
\description{
Deprecated compatibility wrapper for \code{read_gbk()}.
}
